Set preview to active in tab group on LaTeX toggle
Mirrors the behaviour when you edit the source markdown and have
multiple tabs open in the *preview* group, and the associated preview
tab is raised to the top of the group.

In the use case where a preview pane is closed and subsequently reopened
this caused unexpected behaviour; when LaTeX was toggled with focus on
the editor a second preview pane was created. If you continue to close
and open preview panes, when you toggle LaTeX you will get additional
preview panes created.

The source of the problem is that when a preview pane is closed, the
associated `MarkdownPreviewView` object is *destroyed* but not deleted.
This means that the object is still there to listen to the
`markdown-preview-plus:toggle-render-latex` events from
`atom-workspace`. The @editor for each of these objects is the same,
so the condition on L127 (less the `and not @destoryed`) will evaluate
as true, and you end up with multiple objects trying to set their
preview pane as active. All but at most one of these objects will
actually have an open preview tab, and so they will create one.

The addition of the @destroyed flag that is toggled to ture when the
preview pane is destroyed ensures that condition evaluate to false when
it has been destroyed and no extra preview panes will be created.