Preview isnt required to update for restored imgs
Restored images may be either images that we're previously renamed or
previously deleted. To force a restored image to preview a rerender
through either a preview toggle or source edit is required.

Note that on windows pathwatcher does still watch files even after their
their path no longer exists, updating the preview even for restored
images withour a forced rerender, but on OSX it does not.