Fix syncing source lists with preview
The source maps for elements range 1 line longer than the actual source
(which is most likely due to absoring a single `\n` charachter trailing
a block or inline). This causes a problem with tight lists where each
list item is only one line, the sync method would identify the correct
list item to be the one before in the preview if it exists.

Minor changes to method names and return values were also implemented. A
bubble method for token paths was also added to limit the depth of token
paths (currently only conditioning on tables).