Update spec for failing toggle preview activation
This spec only sometimes fails and seems to be that the second preview
pane that the spec was supposed to create was not created in time for
the subsequent `runs` call.

Instead of an `expects` of 1 on the active pane item index, the spec
has been refactor to wait for the index to be 1.

The specs have not failed locally since the update, however after
several failed spec runs before the update, there were several
successfull spec runs before the update.

In any case this update is a reasonable candidate to fix this
psudo-failing spec.